Implement intersection
In order to implement intersection, we could have gone with nested iterated loops over the list of ranges but the time that'd take very quickly becomes unmanageable. Instead, we use an interval tree to avoid the naive iteration cost. We found a gem that implements interval trees well, so we decided to use it.

However, the performance of the `augmented_interval_tree` gem as implemented is not usable for large ranges - we have improved it at https://github.com/greensync/interval-tree/tree/improve-search-performance and we will switch to the gem fork once it's published.

Co-Authored-By: Chris Nankervis <chris@nankervis.com>
Co-Authored-By: Thomas van der Pol <tvdp@hey.com>