Drop support for Ruby 1.8.7
The introduction of the `interval_tree`/`fast_interval_tree` gem for performing intersections is problematic for the Travis builds on Ruby 1.8.7, as it only supports back to Ruby 1.9.3.

Ruby 1.8.7 was [end-of-life'd over 7 years ago](https://www.ruby-lang.org/en/news/2013/06/30/we-retire-1-8-7/). Anyone still using it is in trouble.

Even Ruby 2.2 was [EOL'd two and a half years ago, on 2018-03-31](https://www.ruby-lang.org/en/downloads/branches/#content:~:text=Ruby%202.2).